third_party/WebKit/LayoutTests/http/tests/appcache/exceptions-expected.txt

This tests that 'ApplicationCache' methods throw exceptions with reasonable messages.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS cache.update() threw exception InvalidStateError: Failed to execute 'update' on 'ApplicationCache': there is no application cache to update..
PASS cache.swapCache() threw exception InvalidStateError: Failed to execute 'swapCache' on 'ApplicationCache': there is no newer application cache to swap to..
PASS successfullyParsed is true

TEST COMPLETE
